requirements.txt
PyAudio==0.2.14
faster-whisper==1.1.0
pvporcupine==1.9.5
webrtcvad-wheels==2.0.14
halo==0.0.31
torch
torchaudio
scipy==1.14.1
openwakeword>=0.4.0
websockets==14.1
websocket-client==1.8.0
numpy<2.0.0
RealtimeSTT==0.3.93
opencv-python==4.8.1.78
mediapipe==0.10.20
pillow==10.0.0
pandas==2.1.2
scikit-learn==1.3.0
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
tensorflow==2.18.0
eventlet==0.33.1